Feat: pipe 병렬처리 기능 수정
1. pipe 병렬 처리 로직에서의 signal에 의한 강제 종료코드를 반환하도록
   로직추가
2. 해당 함수 및 파일 재정리

resolved #80
