+++ srcs/exec_another.c
@@ -1,18 +1,11 @@
 #include "../includes/minishell.h"
 
-int	get_exit_status(int status)
-{
-	return (((status) & 0xff00) >> 8);
-}
-
-int	child_process(t_cmds *cmds, char **envp)
+void	child_process(t_cmds *cmds, char **envp)
 {
 	pid_t	pid;
-	int		status;
 
 	pid = fork();
 	cmds->pid = pid;
-	status = 0;
 	if (pid == -1)
 		error_excute(*(cmds->cmd), 0, "Fork function error", 1);
 	if (pid == 0)
@@ -26,19 +19,16 @@ int	child_process(t_cmds *cmds, char **envp)
 	{
 		close(cmds->fd[1]);
 		dup2(cmds->fd[0], 0);
-		//waitpid(pid, &status, 0);
+		close(cmds->fd[0]);
 	}
-	return (get_exit_status(status));
 }
 
-int	last_process(t_cmds *cmds, char **envp)
+void	last_process(t_cmds *cmds, char **envp)
 {
 	pid_t	pid;
-	int		status;
 
 	pid = fork();
 	cmds->pid = pid;
-	status = 0;
 	if (pid == -1)
 		error_excute(*(cmds->cmd), 0, "Fork function error", 1);
 	if (pid == 0)
@@ -46,26 +36,19 @@ int	last_process(t_cmds *cmds, char **envp)
 		redirection(cmds);
 		execute_cmd(cmds->cmd, envp);
 	}
-	else if (pid > 0)
-	{
-		//waitpid(pid, &status, 0);
-	}
-	return (get_exit_status(status));
 }
 
-int	exec_another(t_cmds *cmds, char **envp)
+void	exec_another(t_cmds *cmds, char **envp)
 {
-	int	status;
-
-	status = 0;
-	if (pipe(cmds->fd) == -1)
-	{
-		printf("minishell: %s: Pipe function error\n", *(cmds->cmd));
-		return (1);
-	}
 	if (cmds->next != 0)
-		status = child_process(cmds, envp);
+	{
+		if (pipe(cmds->fd) == -1)
+		{
+			printf("minishell: %s: Pipe function error\n", *(cmds->cmd));
+			exit(1);
+		}
+		child_process(cmds, envp);
+	}
 	else
-		status = last_process(cmds, envp);
-	return (status);
+		last_process(cmds, envp);
 }
